change data size to long long int

+++ src/data_structs/MPR_local_patch_structs.h
@@ -16,7 +16,7 @@ struct mpr_local_patch_struct
 
 	unsigned char* buffer;                   /* data buffer */
 
-	int out_file_size;                       /* the size out file */
+	long long int out_file_size;                       /* the size out file */
 
 	float compression_ratio;                 /* compression ratio for each variable */
